DOT output: Continue parsing file paths for output
When outputting the dependency tree to the DOT files, the program
was now smart enough to realize that translation units in the same
directory do not get the full path to the file and would be dupli-
cated in the graph.

This commit is the first to properly document the efforts the past
few commits have been trying to remedy. The only problem that rem-
ains now is to resolve include statements with relative paths.

I have add a Markdown file to help keep track of the ongoing issu-
es, as opposed to openning an issue on GitHub like I should.